Merge codex/resharper-remediation-2026-09-10 into current main.
Preserve the existing Slice C acceptance and tooling work, and retain
the completed pinned ReSharper launcher and regression-gate documentation.

The combined main tree passes Debug/Release builds and all 3,226 tests,
24 smoke tests, six locked replay goldens, and the ReSharper CLI gate.

+++ src/ProjectAegis.Delegation.Tests/SensorToShooter/SensorToShooterProjectionTests.cs
@@ -4,7 +4,6 @@ using ProjectAegis.Delegation.Decision;
 using ProjectAegis.Delegation.Projection;
 using ProjectAegis.Delegation.SensorToShooter;
 using ProjectAegis.Sim.Catalog;
-using ProjectAegis.Sim.Core;
 using ProjectAegis.Sim.Scenario;
 using NUnit.Framework;
 
@@ -57,7 +56,7 @@ public sealed class SensorToShooterProjectionTests
         var log = new DecisionLog();
         log.AppendContactChange(Change(1, "c1", "hostile-1", "Unknown", "Classified"));
 
-        var staleTick = 1UL + (ulong)KillChainContactStateProjection.DefaultStaleThresholdTicks + 1;
+        var staleTick = 1UL + KillChainContactStateProjection.DefaultStaleThresholdTicks + 1;
         var shooters = new FixedShooterSource(
             new SensorToShooterShooterCandidate(
                 "u1",
